streamlit_app_ML_vizualization/requirements.txt
numpy>=1.23.3
pandas>=1.5.0
scikit-learn>=0.18
scikit-image
scipy>=0.15.1
streamlit>=1.5.0
jupyter>=1.0.0
jupyter_contrib_nbextensions
IPython>=8.5.0
notebook>=6.4.12
matplotlib>=3.6.0
seaborn>=0.11.0
plotly>=5.1.0
pillow>=9.3.0
mlxtend>=0.19.0
pytest
pylint
st_pages
